refactor : qr이란 패키지명은 가독성측면에서 좋지 못하므로 제거
