+++ wabi/src/main/kotlin/com/wap/wabi/event/controller/EventController.kt
@@ -3,6 +3,7 @@ package com.wap.wabi.event.controller
 import com.wap.wabi.common.payload.response.Response
 import com.wap.wabi.event.payload.response.Enum.CheckInTableFilter
 import com.wap.wabi.event.service.EventService
+import com.wap.wabi.event.payload.request.CheckInRequest
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -21,6 +22,14 @@ class EventController (
     @GetMapping("/{eventId}/status")
     fun getCheckStatus(@PathVariable("eventId") eventId: Long) : ResponseEntity<Response>{
         val response = Response.ok(data = eventService.getCheckInStatus(eventId = eventId))
+        return ResponseEntity(response, HttpStatus.OK)
+    }
+
+    @GetMapping("/check-in/qr")
+    fun checkIn(@RequestBody checkInRequest : CheckInRequest) : ResponseEntity<Response> {
+        eventService.checkIn(checkInRequest = checkInRequest)
+
+        val response = Response.ok(data = null);
         return ResponseEntity(response, HttpStatus.OK)
     }
 }
