--- wabi/src/main/kotlin/com/wap/wabi/qr/repository/QrQueryDslRepository.java
@@ -1,14 +1,0 @@
-package com.wap.wabi.qr.repository;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class QrQueryDslRepository {
-    private final JPAQueryFactory queryFactory;
-
-    public QrQueryDslRepository(EntityManager em) {
-        this.queryFactory = new JPAQueryFactory(em);
-    }
-}
